# src/docs/swagger.yaml
definitions:
...
  model.CreateShareLinkResponse:
    properties:
      bucket:
        type: string
      expires_at:
        type: integer
      expires_in:
        type: integer
      key:
        type: string
      share_url:
        type: string
      token:
        type: string
    type: object
...
  model.S3Error:
    description: S3-compatible error response returned on failure
    properties:
      code:
        description: Error code identifying the type of error
        example: NoSuchBucket
        type: string
      message:
        description: Human-readable error message
        example: The specified bucket does not exist
        type: string
      requestId:
        description: Unique request identifier for troubleshooting
        example: 4442587FB7D0A2F9
        type: string
      resource:
        description: The resource associated with the error
        example: my-bucket
        type: string
    type: object
info:
  contact:
    name: API Support
  termsOfService: http://swagger.io/terms/
paths:
...
  /{bucket}/{key}:
    delete:
      description: Deletes an object from the specified bucket
      operationId: deleteobject
      parameters:
      - description: Name of the bucket
        in: path
        name: bucket
        required: true
        type: string
      - description: Object key (path)
        in: path
        name: key
        required: true
        type: string
      - description: AWS4-HMAC-SHA256 authorization header (required if S3_AUTH_ENABLED=true)
        in: header
        name: Authorization
        type: string
      - description: Request timestamp in ISO 8601 format (required if S3_AUTH_ENABLED=true)
        in: header
        name: X-Amz-Date
        type: string
      responses:
        "204":
          description: Object deleted successfully
        "403":
          description: Access denied or signature mismatch
          schema:
            $ref: '#/definitions/model.S3Error'
        "404":
          description: Bucket or object not found
          schema:
            $ref: '#/definitions/model.S3Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.S3Error'
      summary: Delete an object
      tags:
      - objects
    get:
      description: Retrieves an object from the specified bucket
      operationId: getobject
      parameters:
      - description: Name of the bucket
        in: path
        name: bucket
        required: true
        type: string
      - description: Object key (path)
        in: path
        name: key
        required: true
        type: string
      - description: AWS4-HMAC-SHA256 authorization header (required if S3_AUTH_ENABLED=true)
        in: header
        name: Authorization
        type: string
      - description: Request timestamp in ISO 8601 format (required if S3_AUTH_ENABLED=true)
        in: header
        name: X-Amz-Date
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Object data
          headers:
            Content-Length:
              description: Size of the object in bytes
              type: string
            Content-Type:
              description: MIME type of the object
              type: string
            ETag:
              description: MD5 hash of the object (quoted)
              type: string
            Last-Modified:
              description: Last modification time (RFC1123 format)
              type: string
          schema:
            type: file
        "404":
          description: Bucket or object not found
          schema:
            $ref: '#/definitions/model.S3Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.S3Error'
      summary: Get an object
      tags:
      - objects
    head:
      description: Retrieves metadata for an object without downloading the object
        data
      operationId: headobject
      parameters:
      - description: Name of the bucket
        in: path
        name: bucket
        required: true
        type: string
      - description: Object key (path)
        in: path
        name: key
        required: true
        type: string
      - description: AWS4-HMAC-SHA256 authorization header (required if S3_AUTH_ENABLED=true)
        in: header
        name: Authorization
        type: string
      - description: Request timestamp in ISO 8601 format (required if S3_AUTH_ENABLED=true)
        in: header
        name: X-Amz-Date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Object metadata retrieved successfully (no body)
          headers:
            Content-Length:
              description: Size of the object in bytes
              type: string
            Content-Type:
              description: MIME type of the object
              type: string
            ETag:
              description: MD5 hash of the object (quoted)
              type: string
            Last-Modified:
              description: Last modification time (RFC1123 format)
              type: string
        "403":
          description: Access denied or signature mismatch
          schema:
            $ref: '#/definitions/model.S3Error'
        "404":
          description: Bucket or object not found
          schema:
            $ref: '#/definitions/model.S3Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.S3Error'
      summary: Retrieve object metadata
      tags:
      - objects
    put:
      consumes:
      - '*/*'
      description: Uploads an object to the specified bucket with the given key
      operationId: putobject
      parameters:
      - description: Name of the bucket
        in: path
        name: bucket
        required: true
        type: string
      - description: Object key (path)
        in: path
        name: key
        required: true
        type: string
      - description: SHA256 hash of the request payload
        in: header
        name: X-Amz-Content-SHA256
        required: true
        type: string
      - description: AWS4-HMAC-SHA256 authorization header (required if S3_AUTH_ENABLED=true)
        in: header
        name: Authorization
        type: string
      - description: Request timestamp in ISO 8601 format (required if S3_AUTH_ENABLED=true)
        in: header
        name: X-Amz-Date
        type: string
      produces:
      - text/xml
      responses:
        "200":
          description: Object uploaded successfully
          headers:
            ETag:
              description: MD5 hash of the uploaded object
              type: string
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/model.S3Error'
        "403":
          description: Access denied or signature mismatch
          schema:
            $ref: '#/definitions/model.S3Error'
        "404":
          description: Bucket not found
          schema:
            $ref: '#/definitions/model.S3Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.S3Error'
      summary: Upload an object
      tags:
      - objects
  /share/{token}:
    delete:
      description: Revokes a public share link by deleting it
      operationId: deletesharelink
      parameters:
      - description: Share link token
        in: path
        name: token
        required: true
        type: string
      - description: AWS4-HMAC-SHA256 authorization header (required if S3_AUTH_ENABLED=true)
        in: header
        name: Authorization
        type: string
      - description: Request timestamp in ISO 8601 format (required if S3_AUTH_ENABLED=true)
        in: header
        name: X-Amz-Date
        type: string
      responses:
        "204":
          description: Share link deleted successfully
        "403":
          description: Access denied or signature mismatch
          schema:
            $ref: '#/definitions/model.S3Error'
        "404":
          description: Share link not found
          schema:
            $ref: '#/definitions/model.S3Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.S3Error'
      summary: Delete a share link
      tags:
      - sharing
    get:
      description: Downloads an object using a public share token without authentication
      operationId: getsharedobject
      parameters:
      - description: Share link token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Object data
          headers:
            Content-Length:
              description: Size of the object in bytes
              type: string
            Content-Type:
              description: MIME type of the object
              type: string
            ETag:
              description: MD5 hash of the object (quoted)
              type: string
            Last-Modified:
              description: Last modification time (RFC1123 format)
              type: string
          schema:
            type: file
        "403":
          description: Share link expired
          schema:
            $ref: '#/definitions/model.S3Error'
        "404":
          description: Share link not found or object deleted
          schema:
            $ref: '#/definitions/model.S3Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.S3Error'
      summary: Download an object via public share link
      tags:
      - sharing
  /share/create/{bucket}/{key}:
    post:
      consumes:
      - application/json
      description: Generates a unique public link that allows downloading an object
        without authentication
      operationId: createsharelink
      parameters:
      - description: Name of the bucket
        in: path
        name: bucket
        required: true
        type: string
      - description: Object key (path)
        in: path
        name: key
        required: true
        type: string
      - default: 0
        description: Expiration time in seconds (0 for no expiration)
        in: query
        name: expires_in
        type: integer
      - description: AWS4-HMAC-SHA256 authorization header (required if S3_AUTH_ENABLED=true)
        in: header
        name: Authorization
        type: string
      - description: Request timestamp in ISO 8601 format (required if S3_AUTH_ENABLED=true)
        in: header
        name: X-Amz-Date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Share link created successfully
          schema:
            $ref: '#/definitions/model.CreateShareLinkResponse'
        "403":
          description: Access denied or signature mismatch
          schema:
            $ref: '#/definitions/model.S3Error'
        "404":
          description: Bucket or object not found
          schema:
            $ref: '#/definitions/model.S3Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.S3Error'
      summary: Create a public share link for an object
      tags:
      - sharing
swagger: "2.0"